// src/abi/RemarginHelper.d.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */

import {
  ethers,
  EventFilter,
  Signer,
  BigNumber,
  BigNumberish,
  PopulatedTransaction,
} from "ethers";
import {
  Contract,
  ContractTransaction,
  Overrides,
  CallOverrides,
} from "@ethersproject/contracts";
import { BytesLike } from "@ethersproject/bytes";
import { Listener, Provider } from "@ethersproject/providers";
import { FunctionFragment, EventFragment, Result } from "@ethersproject/abi";

interface RemarginHelperInterface extends ethers.utils.Interface {
  functions: {
    "remargin(address,uint256,address,uint256,int256)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "remargin",
    values: [string, BigNumberish, string, BigNumberish, BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "remargin", data: BytesLike): Result;

  events: {};
}

export class RemarginHelper extends Contract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  on(event: EventFilter | string, listener: Listener): this;
  once(event: EventFilter | string, listener: Listener): this;
  addListener(eventName: EventFilter | string, listener: Listener): this;
  removeAllListeners(eventName: EventFilter | string): this;
  removeListener(eventName: any, listener: Listener): this;

  interface: RemarginHelperInterface;

  functions: {
    remargin(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<ContractTransaction>;

    "remargin(address,uint256,address,uint256,int256)"(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<ContractTransaction>;
  };

  remargin(
    from: string,
    fromIndex: BigNumberish,
    to: string,
    toIndex: BigNumberish,
    amount: BigNumberish,
    overrides?: Overrides
  ): Promise<ContractTransaction>;

  "remargin(address,uint256,address,uint256,int256)"(
    from: string,
    fromIndex: BigNumberish,
    to: string,
    toIndex: BigNumberish,
    amount: BigNumberish,
    overrides?: Overrides
  ): Promise<ContractTransaction>;

  callStatic: {
    remargin(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: CallOverrides
    ): Promise<void>;

    "remargin(address,uint256,address,uint256,int256)"(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: CallOverrides
    ): Promise<void>;
  };

  filters: {};

  estimateGas: {
    remargin(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<BigNumber>;

    "remargin(address,uint256,address,uint256,int256)"(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    remargin(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<PopulatedTransaction>;

    "remargin(address,uint256,address,uint256,int256)"(
      from: string,
      fromIndex: BigNumberish,
      to: string,
      toIndex: BigNumberish,
      amount: BigNumberish,
      overrides?: Overrides
    ): Promise<PopulatedTransaction>;
  };
}